+++ frontend/src/components/Menu/DropdownMenu.vue
@@ -18,7 +18,7 @@ interface DropdownMenuItem {
 import { Component, Prop, Vue } from 'vue-property-decorator';
 
 @Component
-export default class DropdownList extends Vue {
+export default class DropdownMenu extends Vue {
   @Prop({ required: true, default: [] }) readonly items!: DropdownMenuItem[];
 
   onClick(index) {
